+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * if a filter is specified for a parameter, the value order in `filters` is used
-* `decorate_x` and `decorate_y` functions
+* `decorate`, `decorate_x` and `decorate_y` functions
 
 ### Removed